optimize: update the code quality
