+++ kernel/ipc/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(big_kernel_common PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/pipe.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/pipe_ops.cpp
+)
